Improve package structure
#433
